refactor: refactoring some components and remove unused functions
