+++ package.json
@@ -19,6 +19,7 @@
     "@tailwindcss/vite": "^4.1.11",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-hook-form": "^7.61.1",
     "react-router-dom": "^7.7.0",
     "tailwindcss": "^4.1.11"
   },
