--- src/services/movingOnPage.ts
@@ -1,14 +1,0 @@
-// export async function stepNext(url:string){
-//   try {
-//     const res = await fetch(url);
-//     const data = await res.json();
-//     this.setResponse(data); // сохраняет responseProps
-//   } catch (e) {
-//     console.error(e);
-//   } finally {
-//     this.setIsLoading(false);
-//   }
-// }
-// export async function stepPrev(url:string){
-//
-// }
