--- src/services/navigation.ts
@@ -1,4 +1,0 @@
-// export function navigate(page:number) {
-//
-//   }
-// }
